refactor(directive syntax): Resolve JSDoc conflict
Change re-thrown error into a console.error() call. This prevents a try-catch
in the processing code from catching the re-thrown error in the commandline
handler.

Add support for the ':' and '$' characters to directive names. The ':'
character is usefule for designating namespaces in a directive name and the
'$' character might be useful in the future.

Add support for formatted strings in the context. String values on the context
will have tokens of the form '${contextPropertyName}' recursively expanded.

Resolve an issue of calling the callback directly in the internal asyncMap
library function. Call done() as expected.

Change directive prefix from '@' to '@@' to avoid conflicts with JSDoc tags.

Change format of CHANGELOG to better resemble a git commit message.

BREAKING CHANGE

All directives with prefix '@' must be prefixed with '@@'.