feat(API, context): API improvements
Add `directr()` export. This makes it easier to mimic the config loading
behaviour exhibited by the `directr` command line tool.

Add the `project` property to the top-level context. This makes it possible to
specify file paths in the context that are always relative to the config file
instead of the current working directory.